# apps/orchestrator/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Preview deployment platform API",
        "title": "PreviewRoll API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/v1/workers": {
            "get": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Get list of all registered workers",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "workers"
                ],
                "summary": "List workers",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/worker.Worker"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/workers/{id}": {
            "get": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Returns a single worker",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "workers"
                ],
                "summary": "Get worker by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Worker ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/worker.Worker"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/workers/deregister": {
            "post": {
                "description": "Mark a worker as offline",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "workers"
                ],
                "summary": "Deregister a worker",
                "parameters": [
                    {
                        "description": "Deregister details",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/worker.DeregisterRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/workers/heartbeat": {
            "post": {
                "description": "Update worker load and last seen timestamp",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "workers"
                ],
                "summary": "Worker heartbeat",
                "parameters": [
                    {
                        "description": "Heartbeat data",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/worker.HeartbeatRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/worker.HeartbeatResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/workers/register": {
            "post": {
                "description": "Register a new worker with the orchestrator",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "workers"
                ],
                "summary": "Register a worker",
                "parameters": [
                    {
                        "description": "Worker details",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/worker.RegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/worker.RegisterResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "worker.Capabilities": {
            "type": "object",
            "properties": {
                "features": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "languages": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "platforms": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "worker.DeregisterRequest": {
            "type": "object",
            "properties": {
                "reason": {
                    "type": "string"
                },
                "worker_id": {
                    "type": "integer"
                }
            }
        },
        "worker.HeartbeatRequest": {
            "type": "object",
            "properties": {
                "current_load": {
                    "type": "integer"
                },
                "worker_id": {
                    "type": "integer"
                }
            }
        },
        "worker.HeartbeatResponse": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "integer"
                }
            }
        },
        "worker.RegisterRequest": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "capabilities": {
                    "$ref": "#/definitions/worker.Capabilities"
                },
                "capacity": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "worker.RegisterResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "worker_id": {
                    "type": "integer"
                }
            }
        },
        "worker.State": {
            "type": "string",
            "enum": [
                "online",
                "offline",
                "busy"
            ],
            "x-enum-varnames": [
                "StateOnline",
                "StateOffline",
                "StateBusy"
            ]
        },
        "worker.Worker": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "capabilities": {
                    "$ref": "#/definitions/worker.Capabilities"
                },
                "capacity": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "current_load": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "last_seen_at": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "state": {
                    "$ref": "#/definitions/worker.State"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "CookieAuth": {
            "type": "apiKey",
            "name": "__Host-sid",
            "in": "cookie"
        }
    }
}